Former-commit-id: f3d07756628d23da604db87f7fb9ca2e46e5b060

+++ web/concrete/models/job.php
@@ -36,6 +36,11 @@ class Job extends Object {
 	
 	public function getJobName() {return $this->jName;}
 	public function getJobDescription() {return $this->jDescription;}	
+	public function getJobHandle() {return $this->jHandle;}
+	public function getPackageHandle() {
+		return PackageList::getHandle($this->pkgID);
+	}
+	
 	
 	//==========================================================
 	// JOB MANAGEMENT - do not override anything below this line 
@@ -77,7 +82,19 @@ class Job extends Object {
 		$val = PASSWORD_SALT . ':' . DIRNAME_JOBS;
 		return md5($val);
 	}
-
+	
+	public static function exportList($xml) {
+		$jl = self::getList();
+		if ($jl->numRows() > 0) {
+			$jx = $xml->addChild('jobs');
+			while($r = $jl->FetchRow()) {
+				$j = Job::getByID($r['jID']);
+				$ch = $jx->addChild('job');
+				$ch->addAttribute('handle',$j->getJobHandle());
+				$ch->addAttribute('package',$j->getPackageHandle());
+			}
+		}	
+	}
 
 	// Job Retrieval 
 	// ==============
@@ -88,6 +105,11 @@ class Job extends Object {
 		$q = "SELECT * FROM Jobs ORDER BY jDateLastRun";
 		$rs = $db->query($q, $v);
 		return $rs;
+	}
+	
+	public static function resetRunningJobs() {
+		$db = Loader::db();
+		$db->Execute('update Jobs set jStatus = \'ENABLED\' where jStatus = \'RUNNING\'');
 	}
 	
 	final static function getByID( $jID=0 ){
@@ -218,7 +240,7 @@ class Job extends Object {
 	}
 	
 	final function executeJob(){
-		
+		Events::fire('on_before_job_execute', $this);
 		$db = Loader::db();
 		$timestampH =date('Y-m-d g:i:s A');
 		$timestamp=date('Y-m-d H:i:s');
@@ -242,7 +264,7 @@ class Job extends Object {
 			$enum = $this->getError();
 		}
 		$rs = $db->query( "INSERT INTO JobsLog (jID, jlMessage, jlTimestamp, jlError) VALUES(?,?,?,?)", array( $this->jID, $resultMsg, $timestamp, $enum ) );
-		
+		Events::fire('on_job_execute', $this);
 		
 		return $resultMsg;
 	}
@@ -283,12 +305,13 @@ class Job extends Object {
 			$db = Loader::db();
 			$db->Execute('insert into Jobs (jName, jDescription, jDateInstalled, jNotUninstallable, jHandle, pkgID) values (?, ?, ?, ?, ?, ?)', 
 				array($j->getJobName(), $j->getJobDescription(), Loader::helper('date')->getLocalDateTime(), 0, $jHandle, $pkg->getPackageID()));
-			
+			Events::fire('on_job_install' $j);
 			return $j;
 		}
 	}
  
 	final public function install(){
+		
 		$db = Loader::db();
 		$jobExists=$db->getOne( 'SELECT count(*) FROM Jobs WHERE jHandle=?', array($this->jHandle) );
 		$vals=array($this->getJobName(),$this->getJobDescription(),  date('Y-m-d H:i:s'), $this->jNotUninstallable, $this->jHandle);
@@ -297,9 +320,14 @@ class Job extends Object {
 		}else{
 			$db->query('INSERT INTO Jobs (jName, jDescription, jDateInstalled, jNotUninstallable, jHandle) VALUES(?,?,?,?,?)',$vals);
 		}
+		Events::fire('on_job_install' $this);
 	}
  
 	final public function uninstall(){
+		$ret = Events::fire('on_job_uninstall', $this);
+		if($ret < 0) {
+			return $ret;
+		}
 		$db = Loader::db();
 		$db->query( 'DELETE FROM Jobs WHERE jHandle=?', array($this->jHandle) );
 	}
